I don't want that github link in the release
I thought about making it automatically change this file in the build process
but that got way too sneaky and hard. It is not too hard to test for it.

If we stick with this, I think I'll create editors to add and remove it. Then release can
be an executor: remove it, commit, tag, add it again, commit.