Remove DSL, reorg runtime and rug intro
The runtime page was a hodgepodge of material from various sources,
mostly introductory.  That content was cleaned, organized, and moved
to the Rug overview.  The runtime was moved down and now contains
information on the, well, runtime.

Remove reviewer content for now.

Remove DSL content from reference, which was the last place it
existed.  Adjust other Rug reference material as necessary.  Some
content was moved to pages that currently are not published for
cleanup.

Quick clean up.