Enable HTMLProofer in Travis CI build
Add HTMLProofer so everything gets checked during build.  Fixed errors
found.  This included what seems like a bug in mkdocs: if the markdown
files does not end it a newline, the last character of the text is
removed.  It appears mkdocs assumes the last character is a newline
and removes it, presumable to assist in markdown conversion to HTML.

Moved Travis CI build logic to script, adding tagging.