Add repo linking instructions, fix non-links
Add instructions for linking channels and repos in the user guide.
Add links to the repo linking instructions throughout the docs.

Fix non-links in commands doc.

In user guide, fix too many capital letters, minor wordsmithing in
user guide, and puts GitHub scopes in italics, as GitHub does and we
do we elsewhere in the docs.

Closes #202

Address @WillGardella review comments.