Remove the GitHub repo link in the header
Setting the repo_url in mkdocs.yml gives me the "Edit this file" button,
which I want. It also gives me a little GitHub badge in the upper right,
which I do NOT want. It's not useful here and it's very distracting
because it's animated.

There is a concern with this commit: to implement this change to
mkdocs-material template, I had to override the partials/header.html file
and remove the part I don't want.
This could be bad when we upgrade the template. We won't see improvements
to the header.
When we upgrade the template, we should check for changes in that file
and, if there are any, re-apply this change to the latest header.html